pyproject.toml
[tool.poetry]
name = "yumi"
version = "0.1.0"
description = "A Chatbot"
authors = ["barweiss"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flake8 = "^7.0.0"
black = "^24.4.0"
discord-py = "^2.3.2"
python-dotenv = "^1.0.1"
requests = "^2.31.0"
mypy = "^1.9.0"
langchain = {extras = ["full"], version = "^0.1.16"}
langchain-google-genai = "^1.0.2"
langchain-mistralai = "^0.1.2"
langchain-openai = "^0.1.3"
types-requests = "^2.31.0.20240406"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.7.0"
tox = "^4.15.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
ignore_errors = true

[[tool.mypy.overrides]]
module = [
    "discord",
    "discord.ext.commands",
]
ignore_missing_imports = true
